Add UTF16 handling to `ixdtf` (#6577)
This PR adds UTF16 handling to `ixdtf`.

For reference, an issue was recently filed on `temporal_rs`
(boa-dev/temporal#275). It was noted that `ixdtf` needs to support
UTF16.

The major changes occur in `utils/ixdtf/src/core.rs`.

Just to note as well, since this would be a breaking change that adds
the import `ixdtf::Slice`, the records module was moved out of parsers
module and into crate module to clean up imports.