Add explicit defaults to FFI enums where clear-cut. (#6692)
This PR does a couple things.

 - Adds `#[default]` to Rust-side enums where unambiguous (e.g. "Auto")
- Adds `#[diplomat::attr(default)]` to Diplomat-side enums where Rust
has a default
- Adds `#[diplomat::attr(default)]` to *non option* Diplomat-side enums
where there is a somewhat sensible default, documenting the reason. In
the case of returned enums an explicit default is still nice for
stability, and there's no downside of lack of explicitness since our
APIs are not accepting them anyway.
- Adds `#[diplomat::attr(default)]` to CalendarKind::Iso and
TimeZoneVariant since having that default be potentially unstable is a
bigger deal than other defaults.


If a default has been added in FFI without a corresponding comment that
means the Rust type has the same default.

A lot of FFI things still do not have an explicit default: I'm
considering "explicit default" to mean that it is a commitment to that
default only changing with intention. Most property enums do not have
one, but they all have a 0 variant which is unlikely to change. A fair
number of our options do not have a default either.

Note that the C++ backend will currently generate a default for every
enum (picking, in order: explicit default, 0 variant, or first variant)
because deferred initialization is a UB risk without it. Us setting
these defaults just provides some stability.

<!--
Thank you for your pull request to ICU4X!

Reminder: try to use [Conventional
Comments](https://conventionalcomments.org/) to make comments clearer.

Please see
https://github.com/unicode-org/icu4x/blob/main/CONTRIBUTING.md for
general
information on contributing to ICU4X.
-->

---------

Co-authored-by: Robert Bastian <4706271+robertbastian@users.noreply.github.com>