Detect rearguard TZDB differences (#6749)
This changes datagen to build the TZDB in rearguard, main, and vanguard
formats, and to complain if any of these versions return different
values. The three current cases where this happens are handled in code,
however the rules for this should be upstreamed to CLDR at some point
(ICU would have to evaluate them at runtime).

Based on #6747