Fix several incorrect type inferences in sort key code (#6708)
These variables were int32_t in the C++ code, but during the port their
types got incorrectly inferred as u8, which can cause a panic on
overflow when processing a long string. I adjust them to usize, which is
Rust's idiom.

The narrowings to u8 should all be safe since we're operating with a
value that fits into a u8 at those points. (The same thing is happening
in the C++ too of course, but it's done silently by the compiler.)