Don't fall back from specific non-location to specific location (#6771)
https://github.com/unicode-org/icu4x/pull/5792 deduplicated
specific-non-location values against specific location, but introduced
specific location as a fall back step if specific-non-location fails. I
came up with this in
[CLDR-18074](https://unicode-org.atlassian.net/browse/CLDR-18074) but
now realise that this is flawed (variants don't make much sense outside
a metazone).

This PR does not undo the deduplication, however it only uses the
specific location format if we expect to find a specific non-location
format, i.e. if we are in a metazone. In particular, it will not produce
"Morroco Standard Time" or "Morocco Daylight Time" anymore, (those terms
didn't make any sense).