Move TZDB compatibility data to CLDR patch (#6770)
For now I'm forking the `metaZones.json` file in-repo, hoping I can get
this into CLDR 48.

This adds the ability to add tzdb-abbreviation -> variant mappings to
metazone periods. This can be used if the `isdst` flag is ambiguous
(i.e. reaguard and main don't agree, datagen checks that there are
overrides for these), or if we believe that the modeling as done in TZDB
does not fit our metazone model (for example `Asia/Samarkand` has a
winter where it stayed on DST, which TZDB models as a base offset
change, but we can keep it in its metazone if we model it as a DST
period).

This PR moves the current in-code overrides to the metazone data.

This approach, however, does not allow adding these overrides to periods
where a zone is not in a metazone, so the offsets for Morocco will
change from (+0, +1) (current code override) to (+1, +0). This is
somewhat fine, because the variant doesn't have much meaning outside a
metazone. I say somewhat, because in
[CLDR-18074](https://unicode-org.atlassian.net/browse/CLDR-18074) I
changed LDML to use the Specific Location format as a fallback from the
Specific Non-Location format; I now consider this to have been a mistake
and will try to revert it (I believe ICU4C has not implemented this
anyway).

#6767 uses this mechanism for more cleanups.