Clarity on line about utils releases (#6798)
I got very confused by this line, which I myself had written in #5209
(as a response to the problem solved by #5198).

This is for utils releases that are a part of an ICU4X patch release.

<!--
Thank you for your pull request to ICU4X!

Reminder: try to use [Conventional
Comments](https://conventionalcomments.org/) to make comments clearer.

Please see
https://github.com/unicode-org/icu4x/blob/main/CONTRIBUTING.md for
general
information on contributing to ICU4X.
-->