Add monotonic_year and tests (#6800)
Supersedes #6762

This also fixes the bug that we were bounds checking extended year: You
should be able to do things like `Gregorian::new(year = -100, era =
None, ...)`, but we errored here.


In the process; I discovered that Chinese was _accepting_ related_iso as
extended year, but _producing_ ICU4C extended year. This is
inconsistent.


This serves as another nail in the coffin for any ICU4C-parity guarantee
we had for extended_year, because we now have found multiple cases where
we are not handling it the ICU4C way already.

The name monotonic_year is to be bikeshed.

---------

Co-authored-by: Robert Bastian <4706271+robertbastian@users.noreply.github.com>