zoneinfo64: make offset_idx/offset_at low level APIs that work purely on the transition table (#6872)
Split out of https://github.com/unicode-org/icu4x/pull/6842, requested
by Robert.

Currently, `offset_at` defaults to attempting to resolve the timezone
rule when given an out-of-range offset. This makes it a higher level API
that is attempting to resolve an offset against both the transitions
table and the offset table (which is what `for_timestamp` does). A lower
level API without this behavior where it *only* works with the
transitions table is useful for clarity, and will become more useful
when we implement prev/next transitions (where resolving
`offset_at(len)` as the next rule is not enough, you need to know which
precise date to resolve it as).

This also splits `Offset` (the return value of offset detection) and
`Transition` (the lower level return value of `offset_at`), because the
`since` value isn't useful to callers and will be wasted effort when
working on Rules.

I've spent some time improving the tests here: we now test more locales
by default; and test invariants.

Most of these locale tests don't test anything meaningful yet since
rules aren't implemented.