Move serde code into separate deserialize module (#6875)
I don't really like having a bunch of nontrivial structs stuffed into a
type, it's harder to navigate and it clutters up the main file. I'd like
for the main file to stick to computation.

Made a separate module.

I _didn't_ move out the rule deserialization code, this can be done
later if we want (it's less of a big deal), but I didn't want to step on
toes of #6871