Reduce the size of `Zone` by deferring lookups (#6903)
Alternative to #6902

This defers the lookups in `zones`, `names`, `rules`, and `regions` from
the constructor, only storing the index in the `Zone` struct.

This will cost an additional lookup at the beginning of each zone
operation, but these are fairly lookup-heavy already anyway. It also
saves the lookups for name and region, which are probably only needed
very rarely. In exchange, the size of `Zone` gets significantly
decreased.

Also adds a `RawZone` abstraction which can be GIGO converted to and
from `Zone`, with the invariant being that the same `ZoneInfo64`
instance is used throughout. Users like temporal can combine this with a
static `LazyLock<ZoneInfo64>` to provide a safe wrapper with minimal
size.