zoneinfo64: Add UtcOffsetSeconds, remove icu_time dep (#6914)
We only use `icu_time` for `UtcOffset`. This is a large extra dependency
(V8 currently doesn't pull it in otherwise), so it would be nice to not
need it.

Open to being convinced that we should use this crate anyway. In the
long term V8 will likely be using icu_time anyway. But allowing clients
to have the option seems good.

I see zoneinfo64 as a low level crate; and icu_time as a higher level
one that you might pair with it for a full-fledged timezone
implementation.

zoneinfo64 does depend on `icu_locale_core` for `Region` but that's a
smaller dep tree.