Add RataDie::in_well_behaved_astronomical_range(), use to avoid panics (#6876)
Attempting to address https://github.com/unicode-org/icu4x/issues/4917,
https://github.com/unicode-org/icu4x/issues/5068,
https://issues.chromium.org/issues/439624688.

Related: https://github.com/unicode-org/icu4x/issues/5778

https://github.com/tc39/proposal-temporal/issues/2869 has told us that
we are allowed to break calendar-internal invariants outside of the
useful range of the calendar.

We currently hit debug assertions for some of these dates. This patches
it up by

This is a partial PR, I haven't fixed all of the assertions. We'll
likely have to make the function in calendrical_calculations pub+hidden
(or make a copy in `icu_calendar`) and use it.

I would like feedback on the approach.