Update Diplomat (#6974)
This includes https://github.com/rust-diplomat/diplomat/pull/964, which
is a soundness fix.

This does change the C++ API a little bit for ListFormatter, but it's a
soundness fix so that takes precedence.

Doing this separately so it can be tested and show up in the changelog;
we'll be cutting a Diplomat release soon.

<!--
Thank you for your pull request to ICU4X!

Reminder: try to use [Conventional
Comments](https://conventionalcomments.org/) to make comments clearer.

Please see
https://github.com/unicode-org/icu4x/blob/main/CONTRIBUTING.md for
general
information on contributing to ICU4X.
-->