Re-remove range-check from Gregorian calendars (#6986)
This is a regression from https://github.com/unicode-org/icu4x/pull/6975
and breaks Temporal. I missed it during review because I thought the
`year_check` code only got hit in `from_codes`, not realizing that
`from_codes` was implemented in terms of `from_fields` anyway.

`from_fields` allows for years to be outside their era, and this is
generally considered a good idea because it prevents surprise errors
when eras change in Japanese.


If the year-checking behavior is desired in from_codes we can add a
simple "check if the returned era was the original era" check.


This was briefly discussed in
https://github.com/unicode-org/icu4x/pull/6910#discussion_r2363290327

<!--
Thank you for your pull request to ICU4X!

Reminder: try to use [Conventional
Comments](https://conventionalcomments.org/) to make comments clearer.

Please see
https://github.com/unicode-org/icu4x/blob/main/CONTRIBUTING.md for
general
information on contributing to ICU4X.
-->

---------

Co-authored-by: Robert Bastian <4706271+robertbastian@users.noreply.github.com>
Co-authored-by: Robert Bastian <robertbastian@unicode.org>