Hardcode KASI-derived data (#7008)
The code and data used for fetching this will be pushed up to a separate
(private) Unicode repo once we have one. You can find the cleaned up
source data in
https://gist.github.com/Manishearth/d8c94a7df22a9eacefc4472a5805322e.

I'm imagining that post-1950 data will change or be removed with
https://github.com/unicode-org/icu4x/pull/7006



The initial motivation here was to fix the apparent ground truth
mismatch found in
https://github.com/unicode-org/icu4x/pull/7007/files#r2393049682. Turns
out it was a different problem, and it has been fixed in
https://github.com/unicode-org/icu4x/pull/7013.

We may potentially need the same discussion as #6970 about whether we
care about these pre-1912 dates, since that's the only time this
diverges.