Improve formatting for Etc/GMT+X zones (#7055)
These should not format as "GMT+?", even if the offset is not set, and
it shouldn't be possible to set a wrong offset.

The special-case code could also live in the specific format
implementations in `icu_datetime`, but there it would be duplicated. It
could also live in `GetField<UtcOffset>`, but I like putting it in
`TimeZoneInfo` because then this is considered for equality of
`TimeZoneInfo`, and by extension `ZonedDateTime`.