Add exhaustive tests for calendar round tripping and `from_fields` and fix overflow bugs (#7065)
This tests the `RataDie -> Date<C> -> Date<Iso> -> Date<C> -> RataDie`
round trip for all dates between `-270000-01-01` and `270000-12-31`. It
also asserts that the produced `Date<C>` follows the same ordering as
the associated `RataDie` (some calendars already test this for some
ranges). For this purpose I added the `DateInner: PartialOrd` bound to
`Calendar`. Four calendars, (including `AnyCalendar`, which is why it
can't be `Ord`) were lacking this implementation.

This discovered a bug in `Hijri`, where the start of the year was stored
relative to the approximate `f64` mean synodic year length, not the
exact tabular year length (this was missed in #7057). I changed the
`debug_assertion!`s in the Hijri code to be actual errors, which are
unwrapped in const[^1] for hardcoded data, and can be shown to not
happen at runtime because the only year infos we create are tabular,
which obeys all the invariants (the simulated rules now fall back to
tabular for years where they would break invariants, I couldn't care
less about this calendar). These invariants also let us remove the while
loops added in #7057.

This PR also adds an exhaustive test for `from_fields` over the extended
years `-270000..=270000`, all month codes, days in `1..=100` and ordinal
months in `1..=100`.

Both tests require `release-with-assertions` to run in an appropriate
time, but the round trip tests then finishes in seconds. We should
figure out some way to run at least that one in CI, as it would
supersede a lot of manually written tests for specific calendars and
ranges, and provides valuable coverage.

[^1]: This requires MSRV 1.83, which is more than 6 versions old, so I
can use

---------

Co-authored-by: Shane F. Carr <shane@unicode.org>