Gate Rules-helpers on unstable (#7094)
According to

https://github.com/unicode-org/icu4x/issues/6659#issuecomment-2967547180,
all experimental APIs should be *gated* (not hidden) on unstable. We
have not yet implemented the plan in that PR, but we should try to
follow it for new things behind the unstable feature.

https://github.com/unicode-org/icu4x/pull/7092 put `Rules` behind the
unstable feature, but it made it `doc(hidden)`. For `Rules`
specifically, it needs to be `doc(hidden)` since datetime uses it, but
the other types should be fully gated.

This PR fully gates the YearInfo types, and makes Rules narrowly
doc(hidden). It also adds additional docs that explain the unstableness
of the parameters on LunarChinese and Hijri since the trait docs will
not be visible when the feature is disabled

---------

Co-authored-by: Robert Bastian <4706271+robertbastian@users.noreply.github.com>