Move DateFields::month_code over to being a byte slice (#7116)
Depends on #7115, depends on #7105

Attempt at fixing #7104.

The user experience of the byte slice is a little bit worse if you want
to use MonthCode::new_normal/new_leap since you have to persist those
for a while.

Most people will not be creating them for long, though. In theory we
could produce byte versions of those methods that produce a limited
range of codes. It's more a problem for our tests that need to do things
like "create many codes in a loop".