Add range invariant to `ArithmeticDate` (#7219)
This adds a `VALID_RD_RANGE` invariant that is derived from the
`VALID_YEAR_RANGE`, which itself is changed to +-10e6 (the range which
we exhaustively test).

The `ArithmeticDate::new_unchecked` and `Calendar::from_rata_die`
methods gains a precondition; the latter usually implies the former, and
itself is usually satisfied by `Calendar::to_rata_die` guaranteeing
this. `Date::from_rata_die` clamps to the valid range; this could be
replaced by a fallible method in the future. The only invocation of
`ArithmeticDate::new_unchecked` that *does not* uphold the precondition
is in `add`.

#7076