Always generate fast-mode data for NFD and NFKD tries (#7222)
The intent is to allow the normalizer to assume that these tries are
fast in the baked data mode.

This changeset removes the ability to generate serde-loadable data with
small-mode tries for NFD and NFKD. Per previous discussion, we can add
the ability to generate small tries for these back if there are users of
ICU4X who really need that mode for these tries.

Fixes #6836