Add snapshot test for until; optimize day/week diffing to use RDs (#7308)
Progress on #7077

Should fix the extreme inefficiency found in
https://github.com/unicode-org/icu4x/issues/7077#issuecomment-3639570717

## The problem

`until()` works by, one at a time, adding years until it "surpasses" the
target date, then doing months, then weeks, then days. This has specific
properties around year/month diffing, but it's less interesting at the
week/day level since those diffs just make sense at the R.D. level.

Simultaneously, the "surpasses" algorithm when given week/day diffs
calls "new_balanced"


https://github.com/unicode-org/icu4x/blob/c46d47f1a81b111892932fd422c10c59a58ae189/components/calendar/src/calendar_arithmetic.rs#L550


and new_balanced does its own little "while loop to build up the date"
situation, which makes the behavior extremely quadratic:


https://github.com/unicode-org/icu4x/blob/c46d47f1a81b111892932fd422c10c59a58ae189/components/calendar/src/calendar_arithmetic.rs#L433-L454

This is bad in any situation where we have a large diff that contains
days, but _especially_ bad when attempting to compute a large diff for
`largest_unit = weeks/days` because it will be doing this again and
again for every week/day.

## This PR

This PR adds a small snapshot test that tests all combinations of
various dates in all calendars with all values of `largest_unit`. It is
marked as `#[ignore]` for now because the test is slow; as I do more
performance work we may be able to un-ignore it, or potentially test it
as release-with-assertions on CI.

While doing so it removes HijriAstronomical from the tests: this
calendar is too slow to test like this.

This PR also adds `to_rata_die_inner()` to `DateFieldsResolver` so it
can be used in CalendarArithmetic.

Finally, _when_ `until` is called with a smallest unit of weeks/days,
it'll just do an RD diff and compute the result.

The test goes from taking 31.53s in release mode to 0.52s.

A thing this PR does _not_ attempt is optimizing the day/week case when
`largest_unit = year/month`. The code appears to not be too slow there.
There is a cap of ~62 days for that case, so it can't get _too_ slow
anyway.

Unfortunately, we can't "just" take the partial diff from the
years/months computation and then do an R.D. computation since it won't
necessarily be the case that that diff applies cleanly. I need to think
about this more, but this PR contains a clear performance win already,
and it's a good start.