Remove `Hijri<AstronomicalSimulation>` logic (#7301)
This reduces the code size of `date_try_from_fields.wasm` from 47kB to
19kB.

We still include 250 years of hardcoded data, and fall back to the
tabular algorithm outside that range.

This calendar specifically documents that we can change the range of the
astronomical calculations (which are incorrect *anyway*).

This is not the size increase observed in
https://github.com/unicode-org/icu4x/issues/7201, as Temporal excludes
this calendar.